+++ src/jestWatchMode.test.ts
@@ -6,22 +6,26 @@ import { getJestWatchMode } from "./jestWatchMode";
   },
   workspace: {
     relativizePath: () => __dirname,
+    config: {
+      get: () => "inherit",
+    },
+    // `contains` used to check presence of a `.git` folder, ensuring a Git project
+    contains: () => true,
   },
 });
 
-describe("watch mode setting", () => {
-  it("uses watchAll mode if option is disabled", () => {
-    (nova as any).config.get = () => false;
+describe("'watch' mode setting", () => {
+  it("uses 'watchAll' mode if option is disabled", () => {
     expect(getJestWatchMode()).toBe("--watchAll");
   });
 
-  it("uses watchAll mode if option is enabled but project is not a git repository", () => {
+  it("uses 'watchAll' mode if option is enabled but project is not a git repository", () => {
     (nova as any).config.get = () => true;
     (nova as any).workspace.contains = () => false;
     expect(getJestWatchMode()).toBe("--watchAll");
   });
 
-  it("uses watchAll mode if project is a git repository but option is disabled", () => {
+  it("uses 'watchAll' mode if project is a git repository but option is disabled", () => {
     (nova as any).config.get = () => false;
     (nova as any).workspace.contains = () => true;
     expect(getJestWatchMode()).toBe("--watchAll");
@@ -31,5 +35,35 @@ describe("watch mode setting", () => {
     (nova as any).config.get = () => true;
     (nova as any).workspace.contains = () => true;
     expect(getJestWatchMode()).toBe("--watch");
+  });
+
+  describe("workspace option overrides global option", () => {
+    it("inherit option - disabled", () => {
+      (nova as any).config.get = () => false;
+      (nova as any).workspace.contains = () => true;
+      (nova as any).workspace.config.get = () => "inherit";
+      expect(getJestWatchMode()).toBe("--watchAll");
+    });
+
+    it("inherit option - enabled", () => {
+      (nova as any).config.get = () => true;
+      (nova as any).workspace.contains = () => true;
+      (nova as any).workspace.config.get = () => "inherit";
+      expect(getJestWatchMode()).toBe("--watch");
+    });
+
+    it("overrides global option - disabled", () => {
+      (nova as any).config.get = () => false;
+      (nova as any).workspace.contains = () => true;
+      (nova as any).workspace.config.get = () => "true";
+      expect(getJestWatchMode()).toBe("--watch");
+    });
+
+    it("overrides global option - enabled", () => {
+      (nova as any).config.get = () => true;
+      (nova as any).workspace.contains = () => true;
+      (nova as any).workspace.config.get = () => "false";
+      expect(getJestWatchMode()).toBe("--watchAll");
+    });
   });
 });